refactor project structure.
