+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/GraphSerializable.scala
@@ -1,0 +1,38 @@
+package com.kakao.s2graph.core.storage
+
+import com.kakao.s2graph.core.types.{InnerValLikeWithTs, InnerValLike}
+import org.apache.hadoop.hbase.util.Bytes
+
+
+trait GraphSerializable {
+  /** serializer */
+  def propsToBytes(props: Seq[(Byte, InnerValLike)]): Array[Byte] = {
+    val len = props.length
+    assert(len < Byte.MaxValue)
+    var bytes = Array.fill(1)(len.toByte)
+    for ((k, v) <- props) bytes = Bytes.add(bytes, v.bytes)
+    bytes
+  }
+
+  def propsToKeyValues(props: Seq[(Byte, InnerValLike)]): Array[Byte] = {
+    val len = props.length
+    assert(len < Byte.MaxValue)
+    var bytes = Array.fill(1)(len.toByte)
+    for ((k, v) <- props) bytes = Bytes.add(bytes, Array.fill(1)(k), v.bytes)
+    bytes
+  }
+
+  def propsToKeyValuesWithTs(props: Seq[(Byte, InnerValLikeWithTs)]): Array[Byte] = {
+    val len = props.length
+    assert(len < Byte.MaxValue)
+    var bytes = Array.fill(1)(len.toByte)
+    for ((k, v) <- props) bytes = Bytes.add(bytes, Array.fill(1)(k), v.bytes)
+    bytes
+  }
+
+  def labelOrderSeqWithIsInverted(labelOrderSeq: Byte, isInverted: Boolean): Array[Byte] = {
+    assert(labelOrderSeq < (1 << 6))
+    val byte = labelOrderSeq << 1 | (if (isInverted) 1 else 0)
+    Array.fill(1)(byte.toByte)
+  }
+}
