+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/hbase/HBaseStorage.scala
@@ -1,0 +1,7 @@
+package com.kakao.s2graph.core.storage.hbase
+object HBaseStorage {
+
+}
+class HBaseStorage {
+
+}
