+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/hbase/HGStorageSerializable.scala
@@ -1,0 +1,7 @@
+package com.kakao.s2graph.core.storage.hbase
+
+import com.kakao.s2graph.core.storage.GKeyValue
+
+trait HGStorageSerializable[E] {
+  def toKeyValues: Seq[GKeyValue]
+}
